Address review feedback: shared padding, canonical, deduplicate SSR
- Move bottom padding (170px + safe-area) into .content CSS class,
  remove hardcoded inline sx from 3 components
- Add canonical URL to fallback metadata branch in profile page
- Extract fetchProfileStatsData() shared server helper, deduplicate
  identical tick-fetching code from you/page.tsx, profile/page.tsx,
  and statistics/page.tsx
- Verified: board tables are truly unified (getBoardTables always
  returns same UNIFIED_TABLES regardless of board name argument)
