+++ packages/web/app/profile/[user_id]/components/profile-sub-page-layout.tsx
@@ -23,7 +23,7 @@ export default function ProfileSubPageLayout({ userId, title, children }: Profil
           {title}
         </Typography>
       </Box>
-      <Box component="main" className={styles.content} sx={{ pb: 'calc(170px + env(safe-area-inset-bottom, 0px))' }}>
+      <Box component="main" className={styles.content}>
         {children}
       </Box>
     </Box>
