+++ packages/web/app/profile/[user_id]/statistics/page.tsx
@@ -1,12 +1,9 @@
 import React from 'react';
 import { Metadata } from 'next';
-import { getServerAuthToken } from '@/app/lib/auth/server-auth';
 import { getServerSession } from 'next-auth/next';
 import { authOptions } from '@/app/lib/auth/auth-options';
-import { cachedUserProfileStats, cachedUserTicks } from '@/app/lib/graphql/server-cached-client';
-import { SUPPORTED_BOARDS } from '@/app/lib/board-data';
+import { fetchProfileStatsData } from '../server-profile-stats';
 import AnalyticsContent from './analytics-content';
-import type { LogbookEntry } from '../utils/profile-constants';
 
 type PageProps = { params: Promise<{ user_id: string }> };
 
@@ -15,46 +12,19 @@ export const metadata: Metadata = {
   robots: { index: false, follow: true },
 };
 
-export default async function ProfileAnalyticsPage({ params }: PageProps) {
+export default async function ProfileStatisticsPage({ params }: PageProps) {
   const { user_id } = await params;
-  const authToken = await getServerAuthToken();
-  let viewerUserId: string | undefined;
-  if (authToken) {
-    const session = await getServerSession(authOptions);
-    viewerUserId = session?.user?.id;
-  }
+  const session = await getServerSession(authOptions);
+  const viewerUserId = session?.user?.id;
 
-  const [initialProfileStats, ...ticksResults] = await Promise.all([
-    cachedUserProfileStats(user_id),
-    ...SUPPORTED_BOARDS.map((boardType) => cachedUserTicks(user_id, boardType)),
-  ]);
-
-  const initialAllBoardsTicks: Record<string, LogbookEntry[]> = {};
-  SUPPORTED_BOARDS.forEach((bt, i) => {
-    const ticks = ticksResults[i];
-    initialAllBoardsTicks[bt] = ticks
-      ? ticks.map((tick) => ({
-          climbed_at: tick.climbedAt,
-          difficulty: tick.difficulty,
-          tries: tick.attemptCount,
-          angle: tick.angle,
-          status: tick.status as LogbookEntry['status'],
-          layoutId: tick.layoutId,
-          boardType: bt,
-          climbUuid: tick.climbUuid,
-        }))
-      : [];
-  });
-
-  const defaultBoard = 'kilter';
-  const initialLogbook = initialAllBoardsTicks[defaultBoard] ?? [];
+  const statsData = await fetchProfileStatsData(user_id);
 
   return (
     <AnalyticsContent
       userId={user_id}
-      initialProfileStats={initialProfileStats}
-      initialAllBoardsTicks={initialAllBoardsTicks}
-      initialLogbook={initialLogbook}
+      initialProfileStats={statsData.initialProfileStats}
+      initialAllBoardsTicks={statsData.initialAllBoardsTicks}
+      initialLogbook={statsData.initialLogbook}
       initialIsOwnProfile={viewerUserId === user_id}
     />
   );
